requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
celery==5.4.0
redis==5.0.8
requests==2.32.3
beautifulsoup4==4.12.3
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-dotenv==1.0.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.35.23
